requirements.txt
django==1.11.2
djangorestframework==3.7.0
django-push-notifications==1.5.0
celery==4.1.0
redis==2.10.6
